refactored the semanticsearch module introducing a tagGeneratorService, queryExpansionSerivce and new ExplainService

+++ molgenis-data-semanticsearch/src/main/java/org/molgenis/data/semanticsearch/explain/bean/ExplainedQueryString.java
@@ -1,7 +1,8 @@
 package org.molgenis.data.semanticsearch.explain.bean;
 
-import com.google.auto.value.AutoValue;
 import org.molgenis.gson.AutoGson;
+
+import com.google.auto.value.AutoValue;
 
 @AutoValue
 @AutoGson(autoValueClass = AutoValue_ExplainedQueryString.class)
@@ -13,9 +14,9 @@ public abstract class ExplainedQueryString
 
 	public abstract String getTagName();
 
-	public abstract double getScore();
+	public abstract float getScore();
 
-	public static ExplainedQueryString create(String matchedWords, String queryString, String tagName, double score)
+	public static ExplainedQueryString create(String matchedWords, String queryString, String tagName, float score)
 	{
 		return new AutoValue_ExplainedQueryString(matchedWords, queryString, tagName, score);
 	}
